PYTHON-5805 Move KMS connect callback tests to test_kms_connect.py
