+++ test/test_encryption.py
@@ -261,493 +261,7 @@ class TestAutoEncryptionOpts(PyMongoTestCase):
             context.host = "evil.example.com"  # type: ignore[misc]
 
 
-class TestKmsConnectCallbackUnit(PyMongoTestCase):
-    """Contract checks for kms_connect_callback that need no KMS server."""
-
-    @staticmethod
-    def _pool_options():
-        return PoolOptions(connect_timeout=10, socket_timeout=10, ssl_context=None)
-
-    def test_non_socket_return_raises_configuration_error(self):
-        def callback(context):
-            return "not-a-socket"
-
-        with self.assertRaisesRegex(ConfigurationError, "must return a connected"):
-            _connect_kms(("kms.example.com", 443), self._pool_options(), callback, 10.0)
-
-    def test_already_wrapped_socket_is_rejected(self):
-        # ssl.SSLSocket passes isinstance but cannot be TLS-wrapped again.
-        ctx = ssl.SSLContext(ssl.PROTOCOL_TLS_CLIENT)
-        ctx.check_hostname = False
-        ctx.verify_mode = ssl.CERT_NONE
-        left, right = socket.socketpair()
-        self.addCleanup(right.close)
-        # No peer needed to produce a genuine ssl.SSLSocket.
-        wrapped = ctx.wrap_socket(left, do_handshake_on_connect=False, server_hostname="x")
-        self.addCleanup(wrapped.close)
-
-        def callback(context):
-            return wrapped
-
-        with self.assertRaisesRegex(ConfigurationError, "unwrapped"):
-            _connect_kms(("kms.example.com", 443), self._pool_options(), callback, 10.0)
-
-    def test_context_receives_host_port_and_timeout(self):
-        received = []
-        left, right = socket.socketpair()
-        self.addCleanup(left.close)
-        self.addCleanup(right.close)
-
-        def callback(context):
-            received.append(context)
-            return left
-
-        # ssl_context=None returns the socket unchanged, so a plain socket is accepted.
-        conn = _connect_kms(("kms.example.com", 443), self._pool_options(), callback, 12.5)
-        self.assertIs(conn, left)
-
-        self.assertEqual(len(received), 1)
-        self.assertEqual(received[0].host, "kms.example.com")
-        self.assertEqual(received[0].port, 443)
-        self.assertEqual(received[0].timeout, 12.5)
-
-    def test_non_blocking_socket_from_callback_is_accepted(self):
-        # Without the driver normalizing the mode, this raises ValueError.
-        server_ctx = ssl.SSLContext(ssl.PROTOCOL_TLS_SERVER)
-        server_ctx.load_cert_chain(CLIENT_PEM)
-        listener = socket.socket()
-        listener.bind(("127.0.0.1", 0))
-        listener.listen(1)
-        self.addCleanup(listener.close)
-
-        def serve():
-            try:
-                conn, _ = listener.accept()
-                server_ctx.wrap_socket(conn, server_side=True).close()
-            except OSError:
-                pass
-
-        threading.Thread(target=serve, daemon=True).start()
-
-        # Built as the driver does, for the flavor-correct type; the local cert won't verify.
-        client_ctx = get_ssl_context(None, None, None, None, True, True, False, _IS_SYNC)
-        options = PoolOptions(connect_timeout=10, socket_timeout=10, ssl_context=client_ctx)
-
-        def connect():
-            sock = socket.create_connection(listener.getsockname(), timeout=10)
-            sock.setblocking(False)
-            return sock
-
-        def callback(context):
-            if _IS_SYNC:
-                return connect()
-            return asyncio.get_running_loop().run_in_executor(None, connect)
-
-        conn = _connect_kms(listener.getsockname(), options, callback, 10.0)
-        self.addCleanup(conn.close)
-        self.assertIsNotNone(conn.gettimeout())
-
-    def test_asyncio_transport_socket_is_rejected(self):
-        # get_extra_info("socket") is a TransportSocket, not a socket.socket.
-        left, right = socket.socketpair()
-        self.addCleanup(left.close)
-        self.addCleanup(right.close)
-
-        def callback(context):
-            return TransportSocket(left)
-
-        with self.assertRaisesRegex(ConfigurationError, "TransportSocket"):
-            _connect_kms(("kms.example.com", 443), self._pool_options(), callback, 10.0)
-
-    def test_http_proxy_helper_tunnels_and_reports_refusal(self):
-        # Covers the CONNECT handshake without KMS credentials.
-        accepted = []
-
-        def stub(listener, reply):
-            try:
-                conn, _ = listener.accept()
-            except OSError:
-                return
-            accepted.append(conn.recv(4096))
-            conn.sendall(reply)
-            conn.close()
-
-        def run_stub(reply):
-            listener = socket.socket()
-            listener.bind(("127.0.0.1", 0))
-            listener.listen(1)
-            self.addCleanup(listener.close)
-            threading.Thread(target=stub, args=(listener, reply), daemon=True).start()
-            return listener.getsockname()
-
-        host, port = run_stub(b"HTTP/1.1 200 Connection Established\r\n\r\n")
-        callback = HTTPProxyKMSConnect(host, port)
-        context = KMSConnectContext(host="kms.example.com", port=443, timeout=10)
-        sock = callback(context)
-        self.addCleanup(sock.close)
-        self.assertIsInstance(sock, socket.socket)
-        self.assertEqual(accepted[0].split(b"\r\n")[0], b"CONNECT kms.example.com:443 HTTP/1.1")
-
-        host, port = run_stub(b"HTTP/1.1 407 Proxy Authentication Required\r\n\r\n")
-        with self.assertRaisesRegex(OSError, "refused CONNECT"):
-            HTTPProxyKMSConnect(host, port)(context)
-
-    def test_tls_proxy_helper_bridges_the_tunnel(self):
-        # Covers the TLS-proxy path and the socketpair relay without KMS creds.
-        server_ctx = ssl.SSLContext(ssl.PROTOCOL_TLS_SERVER)
-        server_ctx.load_cert_chain(CLIENT_PEM)
-        listener = socket.socket()
-        listener.bind(("127.0.0.1", 0))
-        listener.listen(1)
-        self.addCleanup(listener.close)
-
-        def stub_proxy():
-            conn = None
-            try:
-                conn, _ = listener.accept()
-                tls = server_ctx.wrap_socket(conn, server_side=True)
-                request = b""
-                while b"\r\n\r\n" not in request:
-                    chunk = tls.recv(4096)
-                    if not chunk:
-                        return
-                    request += chunk
-                tls.sendall(b"HTTP/1.1 200 Connection Established\r\n\r\n")
-                # The tunnelled peer speaks only after the client does, as a
-                # TLS server would.
-                tls.sendall(b"echo:" + tls.recv(64))
-                tls.close()
-            except OSError:
-                pass
-            finally:
-                if conn is not None:
-                    conn.close()
-
-        threading.Thread(target=stub_proxy, daemon=True).start()
-
-        client_ctx = ssl.SSLContext(ssl.PROTOCOL_TLS_CLIENT)
-        client_ctx.check_hostname = False
-        client_ctx.verify_mode = ssl.CERT_NONE
-        host, port = listener.getsockname()
-        context = KMSConnectContext(host="kms.example.com", port=443, timeout=10)
-
-        sock = HTTPProxyKMSConnect(host, port, client_ctx)(context)
-        self.addCleanup(sock.close)
-        sock.settimeout(10)
-        sock.sendall(b"ping")
-        self.assertEqual(sock.recv(64), b"echo:ping")
-
-    def test_bridge_does_not_inherit_the_connect_deadline(self):
-        # The CONNECT deadline can be much shorter than the KMS request that
-        # follows through the tunnel; the relay must outlast it.
-        server_ctx = ssl.SSLContext(ssl.PROTOCOL_TLS_SERVER)
-        server_ctx.load_cert_chain(CLIENT_PEM)
-        listener = socket.socket()
-        listener.bind(("127.0.0.1", 0))
-        listener.listen(1)
-        self.addCleanup(listener.close)
-
-        def stub_proxy():
-            conn = None
-            try:
-                conn, _ = listener.accept()
-                tls = server_ctx.wrap_socket(conn, server_side=True)
-                request = b""
-                while b"\r\n\r\n" not in request:
-                    chunk = tls.recv(4096)
-                    if not chunk:
-                        return
-                    request += chunk
-                tls.sendall(b"HTTP/1.1 200 Connection Established\r\n\r\n")
-                # Outlast the connect-phase deadline before the KMS side speaks.
-                time.sleep(0.4)
-                tls.sendall(b"echo:" + tls.recv(64))
-                tls.close()
-            except OSError:
-                pass
-            finally:
-                if conn is not None:
-                    conn.close()
-
-        threading.Thread(target=stub_proxy, daemon=True).start()
-
-        client_ctx = ssl.SSLContext(ssl.PROTOCOL_TLS_CLIENT)
-        client_ctx.check_hostname = False
-        client_ctx.verify_mode = ssl.CERT_NONE
-        host, port = listener.getsockname()
-        context = KMSConnectContext(host="kms.example.com", port=443, timeout=0.2)
-
-        sock = HTTPProxyKMSConnect(host, port, client_ctx)(context)
-        self.addCleanup(sock.close)
-        sock.settimeout(10)
-        sock.sendall(b"ping")
-        self.assertEqual(sock.recv(64), b"echo:ping")
-
-    def test_non_coroutine_callback_is_rejected(self):
-        # The async API needs a coroutine function; a plain def must not be
-        # awaited and retried.
-        if _IS_SYNC:
-            raise unittest.SkipTest("a regular function is correct for the sync API")
-
-        left, right = socket.socketpair()
-        self.addCleanup(right.close)
-
-        def callback(context):
-            return left
-
-        with self.assertRaisesRegex(ConfigurationError, "coroutine function"):
-            _connect_kms(("kms.example.com", 443), self._pool_options(), callback, 10.0)
-
-    def test_proxy_closing_before_connect_reply_raises(self):
-        listener = socket.socket()
-        listener.bind(("127.0.0.1", 0))
-        listener.listen(1)
-        self.addCleanup(listener.close)
-
-        def stub_proxy():
-            try:
-                conn, _ = listener.accept()
-                # Read the CONNECT request, then hang up without replying.
-                conn.recv(4096)
-                conn.close()
-            except OSError:
-                pass
-
-        threading.Thread(target=stub_proxy, daemon=True).start()
-
-        host, port = listener.getsockname()
-        context = KMSConnectContext(host="kms.example.com", port=443, timeout=10)
-        with self.assertRaisesRegex(OSError, "proxy closed the connection"):
-            HTTPProxyKMSConnect(host, port)(context)
-
-    def test_tunnel_keeps_bytes_sent_with_the_connect_reply(self):
-        # A proxy may coalesce its 200 with tunnelled bytes; reading past the
-        # header would silently drop them.
-        listener = socket.socket()
-        listener.bind(("127.0.0.1", 0))
-        listener.listen(1)
-        self.addCleanup(listener.close)
-
-        def stub_proxy():
-            try:
-                conn, _ = listener.accept()
-                conn.recv(4096)
-                conn.sendall(b"HTTP/1.1 200 Connection Established\r\n\r\nearly-bytes")
-                conn.close()
-            except OSError:
-                pass
-
-        threading.Thread(target=stub_proxy, daemon=True).start()
-
-        host, port = listener.getsockname()
-        context = KMSConnectContext(host="kms.example.com", port=443, timeout=10)
-        sock = HTTPProxyKMSConnect(host, port)(context)
-        self.addCleanup(sock.close)
-        sock.settimeout(10)
-        self.assertEqual(sock.recv(64), b"early-bytes")
-
-    def test_unconnected_socket_from_callback_is_rejected(self):
-        # The contract says connected; an unconnected socket would otherwise
-        # fail later as a transient error and be retried.
-        bare = socket.socket()
-        self.addCleanup(bare.close)
-
-        def callback(context):
-            return bare
-
-        with self.assertRaisesRegex(ConfigurationError, "already connected"):
-            _connect_kms(("kms.example.com", 443), self._pool_options(), callback, 10.0)
-
-    def test_ipv6_host_is_bracketed_in_connect(self):
-        accepted = []
-        listener = socket.socket()
-        listener.bind(("127.0.0.1", 0))
-        listener.listen(1)
-        self.addCleanup(listener.close)
-
-        def stub_proxy():
-            try:
-                conn, _ = listener.accept()
-                accepted.append(conn.recv(4096))
-                conn.sendall(b"HTTP/1.1 200 Connection Established\r\n\r\n")
-                conn.close()
-            except OSError:
-                pass
-
-        threading.Thread(target=stub_proxy, daemon=True).start()
-
-        host, port = listener.getsockname()
-        context = KMSConnectContext(host="::1", port=443, timeout=10)
-        sock = HTTPProxyKMSConnect(host, port)(context)
-        self.addCleanup(sock.close)
-        self.assertEqual(accepted[0].split(b"\r\n")[0], b"CONNECT [::1]:443 HTTP/1.1")
-
-    def test_oversized_connect_response_is_rejected(self):
-        listener = socket.socket()
-        listener.bind(("127.0.0.1", 0))
-        listener.listen(1)
-        self.addCleanup(listener.close)
-
-        def stub_proxy():
-            conn = None
-            try:
-                conn, _ = listener.accept()
-                conn.recv(4096)
-                # Never sends the terminator.
-                while True:
-                    conn.sendall(b"x" * 1024)
-            except OSError:
-                pass
-            finally:
-                if conn is not None:
-                    conn.close()
-
-        threading.Thread(target=stub_proxy, daemon=True).start()
-
-        host, port = listener.getsockname()
-        context = KMSConnectContext(host="kms.example.com", port=443, timeout=10)
-        with self.assertRaisesRegex(OSError, "oversized CONNECT response"):
-            HTTPProxyKMSConnect(host, port)(context)
-
-    def test_remaining_raises_once_the_deadline_passes(self):
-        from pymongo.encryption_options import _remaining
-
-        self.assertGreater(_remaining(time.monotonic() + 5), 0)
-        with self.assertRaises(socket.timeout):
-            _remaining(time.monotonic() - 1)
-
-    def test_datagram_socket_from_callback_is_rejected(self):
-        # A connected UDP socket passes isinstance and getpeername, but TLS
-        # then raises NotImplementedError, which would be retried.
-        left = socket.socket(socket.AF_INET, socket.SOCK_DGRAM)
-        right = socket.socket(socket.AF_INET, socket.SOCK_DGRAM)
-        self.addCleanup(left.close)
-        self.addCleanup(right.close)
-        right.bind(("127.0.0.1", 0))
-        left.connect(right.getsockname())
-
-        def callback(context):
-            return left
-
-        with self.assertRaisesRegex(ConfigurationError, "stream socket"):
-            _connect_kms(("kms.example.com", 443), self._pool_options(), callback, 10.0)
-
-    def test_kms_request_does_not_retry_a_contract_violation(self):
-        # _connect_kms has no retry loop; the no-retry guarantee is in
-        # kms_request, so exercise that instead.
-        calls = []
-
-        def callback(context):
-            calls.append(context)
-            return "not-a-socket"
-
-        opts = AutoEncryptionOpts({}, "k.d", kms_connect_callback=callback)
-        io = _EncryptionIO(None, mock.MagicMock(), None, opts)
-
-        class StubKmsContext:
-            endpoint = "kms.example.com:443"
-            message = b"request"
-            kms_provider = "aws"
-            usleep = 0
-            bytes_needed = 1
-
-            def feed(self, data):
-                raise AssertionError("should not reach the socket")
-
-            def fail(self):
-                raise AssertionError("a contract violation must not be retried")
-
-        with self.assertRaises(ConfigurationError):
-            io.kms_request(StubKmsContext())
-        self.assertEqual(len(calls), 1)
-
-    def test_contract_violation_surfaces_as_encryption_error(self):
-        # Public operations run under _wrap_encryption_errors, so callers see
-        # EncryptionError with ConfigurationError as its cause.
-        with self.assertRaises(EncryptionError) as caught:
-            with _wrap_encryption_errors():
-                raise ConfigurationError("kms_connect_callback must return ...")
-        self.assertIsInstance(caught.exception.__cause__, ConfigurationError)
-
-    def test_bridge_failure_closes_the_proxy_socket(self):
-        # A failure inside _bridge must not strand the connected proxy socket.
-        listener = socket.socket()
-        listener.bind(("127.0.0.1", 0))
-        listener.listen(1)
-        self.addCleanup(listener.close)
-
-        server_ctx = ssl.SSLContext(ssl.PROTOCOL_TLS_SERVER)
-        server_ctx.load_cert_chain(CLIENT_PEM)
-
-        def stub_proxy():
-            conn = None
-            try:
-                conn, _ = listener.accept()
-                tls = server_ctx.wrap_socket(conn, server_side=True)
-                tls.recv(4096)
-                tls.sendall(b"HTTP/1.1 200 Connection Established\r\n\r\n")
-                tls.close()
-            except OSError:
-                pass
-            finally:
-                if conn is not None:
-                    conn.close()
-
-        threading.Thread(target=stub_proxy, daemon=True).start()
-
-        captured = []
-
-        def failing_bridge(self, proxy):
-            captured.append(proxy)
-            raise OSError("no file descriptors")
-
-        host, port = listener.getsockname()
-        ctx = ssl.SSLContext(ssl.PROTOCOL_TLS_CLIENT)
-        ctx.check_hostname = False
-        ctx.verify_mode = ssl.CERT_NONE
-        context = KMSConnectContext(host="kms.example.com", port=443, timeout=10)
-
-        with mock.patch.object(HTTPProxyKMSConnect, "_bridge", failing_bridge):
-            with self.assertRaisesRegex(OSError, "no file descriptors"):
-                HTTPProxyKMSConnect(host, port, ctx)(context)
-
-        self.assertEqual(captured[0].fileno(), -1, "proxy socket was left open")
-
-    def test_network_error_from_callback_propagates(self):
-        def callback(context):
-            raise OSError("proxy unreachable")
-
-        # Not a ConfigurationError, so kms_request retries it.
-        with self.assertRaises(OSError):
-            _connect_kms(("kms.example.com", 443), self._pool_options(), callback, 10.0)
-
-    @unittest.skipUnless(_HAVE_PYMONGOCRYPT, "pymongocrypt is not installed")
-    def test_client_encryption_accepts_callback(self):
-        def callback(context):
-            raise AssertionError("not called")
-
-        client = self.simple_client()
-        encryption = ClientEncryption(
-            {"local": {"key": b"\x00" * 96}},
-            "keyvault.datakeys",
-            client,
-            OPTS,
-            kms_connect_callback=callback,
-        )
-        self.addCleanup(encryption.close)
-        self.assertIs(encryption._io_callbacks.opts._kms_connect_callback, callback)
-
-    @unittest.skipUnless(_HAVE_PYMONGOCRYPT, "pymongocrypt is not installed")
-    def test_client_encryption_rejects_non_callable(self):
-        client = self.simple_client()
-        with self.assertRaisesRegex(TypeError, "kms_connect_callback must be callable"):
-            ClientEncryption(
-                {"local": {"key": b"\x00" * 96}},
-                "keyvault.datakeys",
-                client,
-                OPTS,
-                kms_connect_callback="not-callable",  # type: ignore[arg-type]
-            )
+# KMS connect callback unit and prose tests moved to test_kms_connect.py.
 
 
 class TestClientOptions(PyMongoTestCase):
@@ -2512,186 +2026,7 @@ class TestKmsTLSProse(EncryptionIntegrationTest):
             self.client_encrypted.create_data_key("aws", master_key=key)
 
 
-KMS_PROXY_HOST = "127.0.0.1"
-KMS_PROXY_PORT = 9004
-KMS_TLS_PROXY_PORT = 9005
-
-AWS_MASTER_KEY = {
-    "region": "us-east-1",
-    "key": "arn:aws:kms:us-east-1:579766882180:key/89fcc2c4-08b0-4bd9-9f25-e30687b580d0",
-}
-
-
-# https://github.com/mongodb/specifications/blob/master/source/client-side-encryption/tests/README.md#kms-connect-callback
-class TestKmsConnectCallbackProse(EncryptionIntegrationTest):
-    @unittest.skipUnless(any(AWS_CREDS.values()), "AWS environment credentials are not set")
-    def setUp(self):
-        super().setUp()
-        self.callback_calls: list[Any] = []
-
-    def plain_callback(self, context):
-        self.callback_calls.append(context)
-        return HTTPProxyKMSConnect(KMS_PROXY_HOST, KMS_PROXY_PORT)(context)
-
-    def tls_callback(self, context):
-        self.callback_calls.append(context)
-        ctx = ssl.create_default_context(cafile=CA_PEM)
-        ctx.check_hostname = False
-        callback = HTTPProxyKMSConnect(KMS_PROXY_HOST, KMS_TLS_PROXY_PORT, ctx)
-        return callback(context)
-
-    def proxy_request(self, method, path, tls=False):
-        """Call the proxy's control endpoints and return the body."""
-        if _IS_SYNC:
-            return self._proxy_request(method, path, tls)
-        return asyncio.get_running_loop().run_in_executor(
-            None, self._proxy_request, method, path, tls
-        )
-
-    def _proxy_request(self, method, path, tls=False):
-        if tls:
-            ctx = ssl.create_default_context(cafile=CA_PEM)
-            ctx.check_hostname = False
-            conn = http.client.HTTPSConnection(
-                f"{KMS_PROXY_HOST}:{KMS_TLS_PROXY_PORT}", context=ctx
-            )
-        else:
-            conn = http.client.HTTPConnection(f"{KMS_PROXY_HOST}:{KMS_PROXY_PORT}")
-        try:
-            conn.request(method, path)
-            return conn.getresponse().read().decode()
-        finally:
-            conn.close()
-
-    def connect_count(self, tls=False):
-        body = self.proxy_request("GET", "/metrics", tls=tls)
-        # One "key value" per line; the server also emits connect_target.
-        for line in body.splitlines():
-            key, _, value = line.partition(" ")
-            if key == "connect_count":
-                return int(value)
-        raise AssertionError(f"no connect_count in metrics body: {body!r}")
-
-    def test_01_plain_http_proxy(self):
-        self.proxy_request("POST", "/reset")
-        encryption = self.create_client_encryption(
-            {"aws": AWS_CREDS},
-            "keyvault.datakeys",
-            self.client,
-            OPTS,
-            kms_connect_callback=self.plain_callback,
-        )
-        encryption.create_data_key("aws", master_key=AWS_MASTER_KEY)
-        self.assertGreaterEqual(self.connect_count(), 1)
-
-    def test_02_https_proxy(self):
-        self.proxy_request("POST", "/reset", tls=True)
-        encryption = self.create_client_encryption(
-            {"aws": AWS_CREDS},
-            "keyvault.datakeys",
-            self.client,
-            OPTS,
-            kms_connect_callback=self.tls_callback,
-        )
-        encryption.create_data_key("aws", master_key=AWS_MASTER_KEY)
-        self.assertGreaterEqual(self.connect_count(tls=True), 1)
-
-    def test_03_auto_encryption_through_proxy(self):
-        self.client.keyvault.datakeys.drop()
-        self.client.db.coll.drop()
-
-        encryption = self.create_client_encryption(
-            {"aws": AWS_CREDS},
-            "keyvault.datakeys",
-            self.client,
-            OPTS,
-            kms_connect_callback=self.plain_callback,
-        )
-        data_key_id = encryption.create_data_key("aws", master_key=AWS_MASTER_KEY)
-        schema = {
-            "bsonType": "object",
-            "properties": {
-                "encrypted_string": {
-                    "encrypt": {
-                        "keyId": [data_key_id],
-                        "bsonType": "string",
-                        "algorithm": "AEAD_AES_256_CBC_HMAC_SHA_512-Deterministic",
-                    }
-                }
-            },
-        }
-
-        self.proxy_request("POST", "/reset")
-        opts = AutoEncryptionOpts(
-            {"aws": AWS_CREDS},
-            "keyvault.datakeys",
-            schema_map={"db.coll": schema},
-            kms_connect_callback=self.plain_callback,
-        )
-        client_encrypted = self.rs_or_single_client(auto_encryption_opts=opts)
-
-        client_encrypted.db.coll.insert_one({"_id": 1, "encrypted_string": "hello"})
-        decrypted = client_encrypted.db.coll.find_one({"_id": 1})
-        self.assertEqual(decrypted["encrypted_string"], "hello")
-
-        raw = self.client.db.coll.find_one({"_id": 1})
-        self.assertIsInstance(raw["encrypted_string"], Binary)
-
-        self.assertGreaterEqual(self.connect_count(), 1)
-
-    def test_04_callback_error(self):
-        def failing_callback(context):
-            raise OSError("proxy is on fire")
-
-        encryption = self.create_client_encryption(
-            {"aws": AWS_CREDS},
-            "keyvault.datakeys",
-            self.client,
-            OPTS,
-            kms_connect_callback=failing_callback,
-        )
-        with self.assertRaisesRegex(EncryptionError, "proxy is on fire"):
-            encryption.create_data_key("aws", master_key=AWS_MASTER_KEY)
-
-    @unittest.skip(
-        "PYTHON-6037 ClientEncryption does not support timeoutMS, so the "
-        "callback always receives the default KMS connect timeout"
-    )
-    def test_05_callback_receives_timeout(self):
-        key_vault_client = self.rs_or_single_client(timeoutMS=1000)
-        encryption = self.create_client_encryption(
-            {"aws": AWS_CREDS},
-            "keyvault.datakeys",
-            key_vault_client,
-            OPTS,
-            kms_connect_callback=self.plain_callback,
-        )
-        encryption.create_data_key("aws", master_key=AWS_MASTER_KEY)
-
-        self.assertTrue(self.callback_calls, "callback was never invoked")
-        for context in self.callback_calls:
-            # Checks only the spec's non-zero requirement, which cannot fail.
-            self.assertIsNotNone(context.timeout)
-            self.assertGreater(context.timeout, 0)
-
-    def test_06_retry_after_network_error(self):
-        state = {"calls": 0}
-
-        def flaky_callback(context):
-            state["calls"] += 1
-            if state["calls"] == 1:
-                raise OSError("first attempt fails")
-            return HTTPProxyKMSConnect(KMS_PROXY_HOST, KMS_PROXY_PORT)(context)
-
-        encryption = self.create_client_encryption(
-            {"aws": AWS_CREDS},
-            "keyvault.datakeys",
-            self.client,
-            OPTS,
-            kms_connect_callback=flaky_callback,
-        )
-        encryption.create_data_key("aws", master_key=AWS_MASTER_KEY)
-        self.assertGreaterEqual(state["calls"], 2)
+# KMS connect callback unit and prose tests moved to test_kms_connect.py.
 
 
 # https://github.com/mongodb/specifications/blob/master/source/client-side-encryption/tests/README.md#kms-tls-options-tests
